map tool

+++ Editor_Window/Engine_SourceCode/tTexture.h
@@ -24,7 +24,7 @@ namespace t::graphics
 		void SetWidth(UINT width) { mWidth = width; }
 		UINT GetHeight() { return mHeight; }
 		void SetHeight(UINT height) { mHeight = height; }
-		void SetSize(Vector2 size) { mWidth = size.x; mHeight = size.y; }
+		void SetSize(Vector2 size) { mWidth = ( UINT ) size.x; mHeight = ( UINT ) size.y; }
 
 		HDC GetHdc() { return mHdc; }
 		eTextureType GetTextureType() { return textureType; }
